requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
executing==2.0.1
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
Markdown==3.5.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.13
